refactor users and posts spec files into 4 files
